feature: refactors the deserializers to enable more code reuse
