+++ vCardLib.Tests/DeserializerTests/V3DeserializerTests.cs
@@ -20,7 +20,7 @@ namespace vCardLib.Tests.DeserializerTests
             vCard vcard = null;
             Assert.DoesNotThrow(delegate
             {
-                vcard = V3Deserializer.Parse(contacts);
+                vcard = V3Deserializer.Parse(contacts, vcard);
             });
             Assert.IsNotNull(vcard);
             //TODO: More assertions to check data integrity should be added
